# artifacts/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KoriAPI",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
...
    "/v1/auth/session": {
      "get": {
        "summary": "Resolve the current session",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "koriAuth": [
              "session"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "summary": "Logout the current session",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "koriAuth": [
              "session"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/auth/mfa/enroll": {
      "post": {
        "summary": "Enroll a TOTP MFA factor",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "koriAuth": [
              "session"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/auth/mfa/verify": {
      "post": {
        "summary": "Verify an enrolled MFA factor",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "koriAuth": [
              "session"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/auth/mfa/disable": {
      "post": {
        "summary": "Disable an MFA factor",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "koriAuth": [
              "session"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/auth/invitations": {
      "get": {
        "summary": "List invitations",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Create a workspace invitation",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/auth/invitations/{id}/accept": {
      "post": {
        "summary": "Accept an invitation token",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "koriAuth": [
              "session"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/service-tokens": {
      "get": {
        "summary": "List service tokens",
        "tags": [
          "service-tokens"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Create a service token",
        "tags": [
          "service-tokens"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/service-tokens/{id}": {
      "delete": {
        "summary": "Revoke a service token",
        "tags": [
          "service-tokens"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/v1/connectors/configs": {
      "get": {
        "summary": "List connector configs",
        "tags": [
          "connectors"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Create or update connector config",
        "tags": [
          "connectors"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/connectors/runs": {
      "get": {
        "summary": "List connector runs",
        "tags": [
          "connectors"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Trigger connector run",
        "tags": [
          "connectors"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "koriAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-kori-session"
      }
    },
...
  }
}